Pull AmazonPurchasingData and AmazonStoreProduct into their own files to match google
